RoutePage redesign init
