+++ src/codeLists/propertyCodeLists.ts
@@ -63,10 +63,10 @@ type IRoutePropertyCodeList = { [key in RouteKeys]: string };
 const routePropertyCodeList: IRoutePropertyCodeList = {
     id: '',
     routePaths: '',
-    routeName: 'REITIN NIMI',
-    routeNameShort: 'REITIN LYHYT NIMI',
-    routeNameSw: 'REITIN NIMI RUOTSIKSI',
-    routeNameShortSw: 'REITIN LYHYT NIMI RUOTSIKSI',
+    routeName: 'NIMI',
+    routeNameShort: 'LYHYT NIMI',
+    routeNameSw: 'NIMI RUOTSIKSI',
+    routeNameShortSw: 'LYHYT NIMI RUOTSIKSI',
     lineId: '',
     line: '',
     modifiedBy: 'MUOKANNUT',
