+++ src/components/sidebar/routeListView/routeList.scss
@@ -5,34 +5,59 @@
     overflow-y: auto;
 
     .routeList {
-        border-top: 1px solid rgb(222, 222, 222);
+        border-top: 1px solid $borderLightGrey;
 
-        .createRoutePathButton {
-            margin-top: 20px;
-            margin-bottom: 10px;
+        .routeListItem {
+            padding-top: 10px;
         }
     }
 
-    input[type='text'] {
-        width: 75%;
-        height: 30px;
-        font-family: hslFont;
-        padding: 3px;
-        border-radius: 5px;
-        border: 1.5px solid rgb(189, 190, 191);
-    }
-
-    .inputTitle {
-        font-size: 0.8rem;
-        font-weight: 900;
-        color: rgb(137, 137, 137);
-        letter-spacing: 0em;
-        padding-bottom: 3px;
-        display: block;
-    }
-
-    .checkboxContainer {
-        padding-top: 10px;
+    .routeName {
         display: flex;
+        align-items: center;
+        overflow: hidden;
+        font-size: 16px;
+        font-weight: normal;
+
+        .routeId {
+            white-space: nowrap;
+        }
+
+        .label {
+            font-size: $largeFontSize;
+            font-weight: bold;
+            margin-left: 10px;
+            margin-right: 10px;
+        }
+
+        .closeView {
+            margin-left: auto;
+        }
+
+        .close {
+            color: $mediumGrey;
+            transition: 0.5s;
+            font-size: 18px;
+            cursor: pointer;
+        }
+
+        .close:hover {
+            color: $darkGrey;
+            transition: 0.5s;
+            font-size: 18px;
+            cursor: pointer;
+        }
+    }
+
+    .routeItemWrapper {
+        border-top: 1.5px solid #bebebe;
+    }
+
+    .buttonContainer {
+        margin-top: 10px;
+        margin-bottom: 10px;
+        div {
+            margin: 0px;
+        }
     }
 }
