+++ src/components/sidebar/routeListView/routePathListTab.scss
@@ -1,7 +1,7 @@
 @import '../../shared/styles/common';
 @import '../../shared/styles/flexbox';
 
-.routeItemView {
+.routePathListTab {
     .groupedRoutes {
         padding: 10px 15px;
 
@@ -18,48 +18,6 @@
 
     .shadow {
         background-color: white;
-    }
-
-    .routeName {
-        display: flex;
-        align-items: center;
-        overflow: hidden;
-        font-size: 16px;
-        font-weight: normal;
-
-        .routeId {
-            cursor: pointer;
-        }
-
-        .routeId:hover {
-            background-color: $linkHoverBackgroundColor;
-            transition: 0.5s;
-        }
-
-        .label {
-            font-size: $largeFontSize;
-            font-weight: bold;
-            margin-left: 10px;
-            margin-right: 10px;
-        }
-
-        .closeView {
-            margin-left: auto;
-        }
-
-        .close {
-            color: $mediumGrey;
-            transition: 0.5s;
-            font-size: 18px;
-            cursor: pointer;
-        }
-
-        .close:hover {
-            color: $darkGrey;
-            transition: 0.5s;
-            font-size: 18px;
-            cursor: pointer;
-        }
     }
 
     .routePathContainer {
