+++ src/components/sidebar/routeListView/routeTab.scss
@@ -2,7 +2,7 @@
 @import '../../shared/styles/flexbox';
 @import '../../shared/styles/form';
 
-.routeInfoTabView {
+.routeTab {
     display: flex;
     flex-direction: column;
     overflow-y: auto;
