+++ src/components/sidebar/routeView/routeForm.scss
@@ -1,0 +1,3 @@
+@import '../../shared/styles/common';
+@import '../../shared/styles/flexbox';
+@import '../../shared/styles/form';
