--- src/components/sidebar/routeView/routePathTab.scss
@@ -1,12 +1,0 @@
-@import '../../shared/styles/common';
-
-.routePathTabView {
-    height: 100%;
-    flex-direction: column;
-    display: flex;
-    overflow-y: auto;
-
-    .content {
-        padding: 15px;
-    }
-}
