+++ src/components/sidebar/routeView/routeView.scss
@@ -1,16 +1,11 @@
 @import '../../shared/styles/common';
-@import '../../shared/styles/sidebar';
 @import '../../shared/styles/flexbox';
 @import '../../shared/styles/form';
 
 .routeView {
-    width: $sidebarWide;
-    height: 100%;
     display: flex;
     flex-direction: column;
-
-    .content {
-        overflow-y: auto;
-        flex-grow: 1;
-    }
+    overflow-y: auto;
+    height: 100%;
+    padding: 15px;
 }
