+++ src/stores/routeStore.ts
@@ -1,14 +1,22 @@
 import _ from 'lodash';
 import { action, computed, observable } from 'mobx';
 import { IRoute } from '~/models';
+import routeValidationModel, {
+    IRouteValidationModel
+} from '~/models/validationModels/routeValidationModel';
+import { IValidationResult } from '~/validation/FormValidator';
+import ValidationStore, { ICustomValidatorMap } from './validationStore';
 
 class RouteStore {
     @observable private _route: IRoute | null;
     @observable private _oldRoute: IRoute | null;
-    @observable private _isEditingDisabled: boolean;
+    @observable private _isNewRoute: boolean;
+    @observable private _existingRouteIds: string[] = [];
+    @observable private _routeIdToEdit: string | null;
+    private _validationStore: ValidationStore<IRoute, IRouteValidationModel>;
 
     constructor() {
-        this._isEditingDisabled = true;
+        this._validationStore = new ValidationStore();
     }
 
     @computed
@@ -22,6 +30,21 @@ class RouteStore {
     }
 
     @computed
+    get isNewRoute(): boolean {
+        return this._isNewRoute;
+    }
+
+    @computed
+    get existingRouteIds(): string[] {
+        return this._existingRouteIds;
+    }
+
+    @computed
+    get routeIdToEdit(): string | null {
+        return this._routeIdToEdit;
+    }
+
+    @computed
     get isDirty() {
         // line and routePaths can't change in routeView, omit them from
         // the comparison to prevent lag
@@ -32,14 +55,51 @@ class RouteStore {
     }
 
     @computed
-    get isEditingDisabled() {
-        return this._isEditingDisabled;
+    get invalidPropertiesMap() {
+        return this._validationStore.getInvalidPropertiesMap();
+    }
+
+    @computed
+    get isRouteFormValid() {
+        return this._validationStore.isValid();
     }
 
     @action
-    public setRoute = (route: IRoute) => {
-        this._route = route;
-        this.setOldRoute(this._route);
+    public init = ({ route, isNewRoute }: { route: IRoute; isNewRoute: boolean }) => {
+        this._route = _.cloneDeep(route);
+        this.setOldRoute(route);
+        this._isNewRoute = isNewRoute;
+        const customValidatorMap: ICustomValidatorMap = {
+            id: {
+                validator: (route: IRoute, property: string, routeId: string) => {
+                    if (!this._isNewRoute) return;
+                    if (Boolean(this._existingRouteIds.includes(routeId))) {
+                        const validationResult: IValidationResult = {
+                            isValid: false,
+                            errorMessage: `Reitti ${routeId} on jo olemassa.`
+                        };
+                        return validationResult;
+                    }
+                    return;
+                }
+            }
+        };
+        this._validationStore.init(this._route, routeValidationModel, customValidatorMap);
+    };
+
+    @action
+    public setRouteToEdit = (route: IRoute | null) => {
+        if (!route) {
+            this._routeIdToEdit = null;
+            return;
+        }
+        if (route.id === this._routeIdToEdit) {
+            this.resetChanges();
+            this._routeIdToEdit = null;
+        } else {
+            this.init({ route, isNewRoute: false });
+            this._routeIdToEdit = route.id;
+        }
     };
 
     @action
@@ -49,31 +109,26 @@ class RouteStore {
 
     @action
     public updateRouteProperty = (property: keyof IRoute, value: string | number | Date) => {
-        this._route = {
-            ...this._route!,
-            [property]: value
-        };
+        (this._route as any)[property] = value;
+        this._validationStore.updateProperty(property, value);
     };
 
     @action
-    public setIsEditingDisabled = (isEditingDisabled: boolean) => {
-        this._isEditingDisabled = isEditingDisabled;
-    };
-
-    @action
-    public toggleIsEditingDisabled = () => {
-        this._isEditingDisabled = !this._isEditingDisabled;
+    public setExistingRouteIds = (existingRouteIds: string[]) => {
+        this._existingRouteIds = existingRouteIds;
     };
 
     @action
     public clear = () => {
         this._route = null;
+        this._oldRoute = null;
+        this._validationStore.clear();
     };
 
     @action
     public resetChanges = () => {
         if (this._oldRoute) {
-            this.setRoute(this._oldRoute);
+            this.init({ route: this._oldRoute, isNewRoute: this._isNewRoute });
         }
     };
 }
